claimcheck: refactor writer/stager separation
- stager owns the full S3/Parquet lifecycle via Batch.finalize, which
  merges parquet flush, pw.Close, and multipart upload completion into
  one coherent method; span is started in stage and ended exactly once
  inside finalize or Cleanup
- writer.go owns Batch.Produce via the produce closure: calls
  batch.finalize, serializes the envelope, writes to Kafka, and
  calls stager.unstage on any post-finalize failure (serialization
  or write)
- stager.unstage takes *Batch and uses s3Factory; no s3Client field
  on Batch
- BatchOption types (batchConfig, WithBatchKey, WithBatchHeaders)
  moved to writer.go where they belong
- Remove abort(), endSpan(), spanEnded, and the old finalizeUpload
  split; Batch.Produce is now a three-line guard + delegate
